Add multiroom functionality
This adds multiroom functionality to the application. this namespaces a
URL to a specific room (http://app.com/foo => foo). Rooms are created
dynamically by the client detecting their own URL. Ensuring a connect on
a certain room. All functions now only broadcast to their respective rooms.

It currently isn't really possible to be in two rooms at once. But this
should be considered later on.